Construct an interactome by comparing bait and control background across experimental conditions

Construct an interactome by comparing bait and control background across experimental conditions

analyse_interactome(Intensity, Intensity_na_replaced = Intensity,
  conditions, bait_gene_name, Npep, Protein.IDs, names, bckg_bait,
  bckg_ctrl, by_conditions = TRUE, pool_background = TRUE,
  log_test = TRUE, log_stoichio = TRUE, substract_ctrl = TRUE,
  use_mean_for_bait = TRUE)

Arguments

Intensity
a data frame of protein intensities (without replacement of missing values). columns are experimental samples and rows are proteins

Intensity_na_replaced
a data frame of protein intensities (with replacement of missing values). columns are experimental samples and rows are proteins

conditions
: data frame describing the conditions corresponding to the columns of the data frame Intensity

bait_gene_name
: The gene name of the bait

Npep
: vector containing the number of theoretically observable peptide per protein (same length as dim(df)[1])

Protein.IDs
: vector containing protein IDs (same length as dim(df)[1])

names
: vector containing protein names (same length as dim(df)[1])

bckg_bait
: Name of the bait background as found in conditions$bckg

bckg_ctrl
: Name of the control background as found in conditions$bckg

by_conditions
option to perform the comparison between bait and control group for each condition

pool_background
option to use all control background conditions as one control group for all conditions

log_test
logical, perform t-test on log transform intensities

log_stoichio
logical, use the geometric mean instead of the arithmetic mean to compute stoichiometries

substract_ctrl
logical, substract ctrl intensities in the calculation of stoichiometries

use_mean_for_bait
logical, average bait intensities across all conditions to compute interaction stoichiometries

Value

an object of class InteRactome, i.e a list including the following elements :

conditions : a vector of experimental conditions.

names : a vector of names (by default gene names are used).

p_val : a list of vectors containing the p values associated to each experimental condition.

fold_change : a list of vectors containing the fold change associated to each experimental condition.

... : other variables.