Compute Element Deflection for an Event

This function calculates the deflection for an Actor, Behavior, Object event. It assumes that the first input is an identity corresponding to the actor, the second, the behavior, and last, the object. Each of these terms must be in the US 2015 dictionary.

element_deflection(
  act,
  beh,
  obj,
  dictionary = "us",
  equation = c("us", "user supplied"),
  eq_df = NULL
)

Arguments

act
lowercase string corresponding to the actor identity

beh
lowercase string corresponding to the behavior term

obj
lowercase string corresponding to the object identity

dictionary
which dictionary to use, currently set to "us"

equation
which equation to use - you can either set it to "us" for the us 1978 equations, or "user supplied")

eq_df
if you select "user supplied" for equation, this parameter should be your equation dataframe, which (should have been reshaped by the reshape_new_equation function prior)

Value

the deflection produced by each element of the event, a 9 x 1 matrix

Examples

element_deflection("ceo", "advise", "benefactor", equation = "us")

#> # A tibble: 9 x 7
#>   term      element dimension fundamental_sentim… trans_imp difference  sqd_diff
#>   <chr>     <chr>   <chr>                   <dbl>     <dbl>      <dbl>     <dbl>
#> 1 ceo       A       E                        0.71    1.51     -0.802     6.44e-1
#> 2 ceo       A       P                        3.22    2.38      0.842     7.10e-1
#> 3 ceo       A       A                        1.48    1.02      0.455     2.07e-1
#> 4 advise    B       E                        2.57    1.60      0.970     9.41e-1
#> 5 advise    B       P                        2.28    2.09      0.194     3.76e-2
#> 6 advise    B       A                        0.28    0.613    -0.333     1.11e-1
#> 7 benefact… O       E                        1.97    1.50      0.470     2.21e-1
#> 8 benefact… O       P                        1.98    1.08      0.903     8.16e-1
#> 9 benefact… O       A                        0.1     0.0975    0.00250   6.25e-6