Identify conditions (background, time of stimulation, biological and technical replicates) from column names

Identify conditions (background, time of stimulation, biological and technical replicates) from column names

identify_conditions(df, Column_intensity_pattern = "^Intensity.",
  split = "_", bckg_pos = 1, bio_pos = 3, time_pos = 2,
  tech_pos = 4)

Arguments

df
A dataframe containing protein intensities. By default, protein intensity column names start by "Intensity." (use parameter Column_intensity_pattern to change)

Column_intensity_pattern
Pattern (regular exrpression) used to identfy df's columns containing protein intensity values

split
Character used to split column names into substrings

bckg_pos
Position of the sample background in splitted column names

bio_pos
Position of the sample biological replicate in splitted column names

time_pos
Position of the sample experimental condition in splitted column names

tech_pos
Position of the sample technical replicate in splitted column names

Value

a data frame describing experimental samples in terms of background, biological and technical replicates, and experimental conditions

Examples

#load data :
data("proteinGroups_Cbl")
# You can identify columns and their description separately using `identify_conditions()`
cond <- identify_conditions(proteinGroups_Cbl)
print.data.frame(cond)

#>                       column bckg time  bio tech
#> 1    Intensity.Cbl_0_Ech1_R1  Cbl    0 Ech1   R1
#> 2    Intensity.Cbl_0_Ech1_R2  Cbl    0 Ech1   R2
#> 3    Intensity.Cbl_0_Ech1_R3  Cbl    0 Ech1   R3
#> 4    Intensity.Cbl_0_Ech2_R1  Cbl    0 Ech2   R1
#> 5    Intensity.Cbl_0_Ech2_R2  Cbl    0 Ech2   R2
#> 6    Intensity.Cbl_0_Ech2_R3  Cbl    0 Ech2   R3
#> 7    Intensity.Cbl_0_Ech3_R1  Cbl    0 Ech3   R1
#> 8    Intensity.Cbl_0_Ech3_R2  Cbl    0 Ech3   R2
#> 9    Intensity.Cbl_0_Ech3_R3  Cbl    0 Ech3   R3
#> 10 Intensity.Cbl_030_Ech1_R1  Cbl  030 Ech1   R1
#> 11 Intensity.Cbl_030_Ech1_R2  Cbl  030 Ech1   R2
#> 12 Intensity.Cbl_030_Ech1_R3  Cbl  030 Ech1   R3
#> 13 Intensity.Cbl_030_Ech2_R1  Cbl  030 Ech2   R1
#> 14 Intensity.Cbl_030_Ech2_R2  Cbl  030 Ech2   R2
#> 15 Intensity.Cbl_030_Ech2_R3  Cbl  030 Ech2   R3
#> 16 Intensity.Cbl_030_Ech3_R1  Cbl  030 Ech3   R1
#> 17 Intensity.Cbl_030_Ech3_R2  Cbl  030 Ech3   R2
#> 18 Intensity.Cbl_030_Ech3_R3  Cbl  030 Ech3   R3
#> 19 Intensity.Cbl_120_Ech1_R1  Cbl  120 Ech1   R1
#> 20 Intensity.Cbl_120_Ech1_R2  Cbl  120 Ech1   R2
#> 21 Intensity.Cbl_120_Ech1_R3  Cbl  120 Ech1   R3
#> 22 Intensity.Cbl_120_Ech2_R1  Cbl  120 Ech2   R1
#> 23 Intensity.Cbl_120_Ech2_R2  Cbl  120 Ech2   R2
#> 24 Intensity.Cbl_120_Ech2_R3  Cbl  120 Ech2   R3
#> 25 Intensity.Cbl_120_Ech3_R1  Cbl  120 Ech3   R1
#> 26 Intensity.Cbl_120_Ech3_R2  Cbl  120 Ech3   R2
#> 27 Intensity.Cbl_120_Ech3_R3  Cbl  120 Ech3   R3
#> 28 Intensity.Cbl_300_Ech1_R1  Cbl  300 Ech1   R1
#> 29 Intensity.Cbl_300_Ech1_R2  Cbl  300 Ech1   R2
#> 30 Intensity.Cbl_300_Ech1_R3  Cbl  300 Ech1   R3
#> 31 Intensity.Cbl_300_Ech2_R1  Cbl  300 Ech2   R1
#> 32 Intensity.Cbl_300_Ech2_R2  Cbl  300 Ech2   R2
#> 33 Intensity.Cbl_300_Ech2_R3  Cbl  300 Ech2   R3
#> 34 Intensity.Cbl_300_Ech3_R1  Cbl  300 Ech3   R1
#> 35 Intensity.Cbl_300_Ech3_R2  Cbl  300 Ech3   R2
#> 36 Intensity.Cbl_300_Ech3_R3  Cbl  300 Ech3   R3
#> 37 Intensity.Cbl_600_Ech1_R1  Cbl  600 Ech1   R1
#> 38 Intensity.Cbl_600_Ech1_R2  Cbl  600 Ech1   R2
#> 39 Intensity.Cbl_600_Ech1_R3  Cbl  600 Ech1   R3
#> 40 Intensity.Cbl_600_Ech2_R1  Cbl  600 Ech2   R1
#> 41 Intensity.Cbl_600_Ech2_R2  Cbl  600 Ech2   R2
#> 42 Intensity.Cbl_600_Ech2_R3  Cbl  600 Ech2   R3
#> 43 Intensity.Cbl_600_Ech3_R1  Cbl  600 Ech3   R1
#> 44 Intensity.Cbl_600_Ech3_R2  Cbl  600 Ech3   R2
#> 45 Intensity.Cbl_600_Ech3_R3  Cbl  600 Ech3   R3
#> 46    Intensity.WT_0_Ech1_R1   WT    0 Ech1   R1
#> 47    Intensity.WT_0_Ech1_R2   WT    0 Ech1   R2
#> 48    Intensity.WT_0_Ech1_R3   WT    0 Ech1   R3
#> 49    Intensity.WT_0_Ech2_R1   WT    0 Ech2   R1
#> 50    Intensity.WT_0_Ech2_R2   WT    0 Ech2   R2
#> 51    Intensity.WT_0_Ech2_R3   WT    0 Ech2   R3
#> 52    Intensity.WT_0_Ech3_R1   WT    0 Ech3   R1
#> 53    Intensity.WT_0_Ech3_R2   WT    0 Ech3   R2
#> 54    Intensity.WT_0_Ech3_R3   WT    0 Ech3   R3
#> 55  Intensity.WT_030_Ech1_R1   WT  030 Ech1   R1
#> 56  Intensity.WT_030_Ech1_R2   WT  030 Ech1   R2
#> 57  Intensity.WT_030_Ech1_R3   WT  030 Ech1   R3
#> 58  Intensity.WT_030_Ech2_R1   WT  030 Ech2   R1
#> 59  Intensity.WT_030_Ech2_R2   WT  030 Ech2   R2
#> 60  Intensity.WT_030_Ech2_R3   WT  030 Ech2   R3
#> 61  Intensity.WT_030_Ech3_R1   WT  030 Ech3   R1
#> 62  Intensity.WT_030_Ech3_R2   WT  030 Ech3   R2
#> 63  Intensity.WT_030_Ech3_R3   WT  030 Ech3   R3
#> 64  Intensity.WT_120_Ech1_R1   WT  120 Ech1   R1
#> 65  Intensity.WT_120_Ech1_R2   WT  120 Ech1   R2
#> 66  Intensity.WT_120_Ech1_R3   WT  120 Ech1   R3
#> 67  Intensity.WT_120_Ech2_R1   WT  120 Ech2   R1
#> 68  Intensity.WT_120_Ech2_R2   WT  120 Ech2   R2
#> 69  Intensity.WT_120_Ech2_R3   WT  120 Ech2   R3
#> 70  Intensity.WT_120_Ech3_R1   WT  120 Ech3   R1
#> 71  Intensity.WT_120_Ech3_R2   WT  120 Ech3   R2
#> 72  Intensity.WT_120_Ech3_R3   WT  120 Ech3   R3
#> 73  Intensity.WT_300_Ech1_R1   WT  300 Ech1   R1
#> 74  Intensity.WT_300_Ech1_R2   WT  300 Ech1   R2
#> 75  Intensity.WT_300_Ech1_R3   WT  300 Ech1   R3
#> 76  Intensity.WT_300_Ech2_R1   WT  300 Ech2   R1
#> 77  Intensity.WT_300_Ech2_R2   WT  300 Ech2   R2
#> 78  Intensity.WT_300_Ech2_R3   WT  300 Ech2   R3
#> 79  Intensity.WT_300_Ech3_R1   WT  300 Ech3   R1
#> 80  Intensity.WT_300_Ech3_R2   WT  300 Ech3   R2
#> 81  Intensity.WT_300_Ech3_R3   WT  300 Ech3   R3
#> 82  Intensity.WT_600_Ech1_R1   WT  600 Ech1   R1
#> 83  Intensity.WT_600_Ech1_R2   WT  600 Ech1   R2
#> 84  Intensity.WT_600_Ech1_R3   WT  600 Ech1   R3
#> 85  Intensity.WT_600_Ech2_R1   WT  600 Ech2   R1
#> 86  Intensity.WT_600_Ech2_R2   WT  600 Ech2   R2
#> 87  Intensity.WT_600_Ech2_R3   WT  600 Ech2   R3
#> 88  Intensity.WT_600_Ech3_R1   WT  600 Ech3   R1
#> 89  Intensity.WT_600_Ech3_R2   WT  600 Ech3   R2
#> 90  Intensity.WT_600_Ech3_R3   WT  600 Ech3   R3

# and use it as parameters for function InteRact()
res <- InteRact(proteinGroups_Cbl, bait_gene_name = "Cbl", condition = cond)

#> Warning: Column 'Score' not available : Data NOT Filtered based on portein identification score

#> Contaminant proteins discarded
#> Proteins with no gene name available discarded
#> Number of theoretically observable peptides unavailable : used MW instead
#> Merge protein groups associated to the same gene name (sum of intensities) 
#> Rescale median intensity across conditions
#> Replace missing values and perform interactome analysis for 1 replicates
#> Nrep=1
#> Averaging 1 interactomes

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced

#> Warning: NaNs produced