maximally_confirm_behavior

Identify the behavior that would maximally confirm the identities of actor and object pairing

Identify the behavior that would maximally confirm the identities of actor and object pairing

maximally_confirm_behavior(
  d,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)

maximally_confirm_behavior(
  d,
  equation_key = NULL,
  equation_gender = NULL,
  eq_df = NULL,
  ...
)

Arguments

d

events dataframe reshaped by reshape_events_dataframe function with an actor and object pairing

equation_key

a string corresponding to the equation key from actdata

equation_gender

either average, male, or female, depending on if you are using gendered equations

eq_df

if you select "user supplied" for equation, this parameter should be your equation dataframe, which (should have been reshaped by the reshape_new_equation function prior)

actor

lowercase string corresponding to the actor identity

beh

lowercase string corresponding to the behavior term

object

lowercase string corresponding to the object identity

gender

either average, male, or female, depending on if you are using gendered equations

dictionary_key

a string corresponding to the dictionary from actdata you are using for cultural EPA measurements

Value

3 digit EPA indicating the optimal behavior

3 digit EPA indicating the optimal behavior

Examples

d <- tibble::tibble(actor = "ceo", object = "benefactor")
d <- reshape_events_df(df = d, df_format = "wide", dictionary_key = "usfullsurveyor2015", dictionary_gender = "average")
#> Joining with `by = join_by(term, component)`
maximally_confirm_behavior(d = d, equation_key = "us2010", equation_gender = "average")
#> # A tibble: 1 × 4
#>   opt_E opt_P opt_A term 
#>   <dbl> <dbl> <dbl> <chr>
#> 1    NA    NA    NA actor