Total Interaction Strength

Calculates proportion of variability of prediction function that comes from interaction effects.

total_interaction(object, ...)

# S3 method for default
total_interaction(
  object,
  v,
  X,
  pred_fun = stats::predict,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

# S3 method for ranger
total_interaction(
  object,
  v,
  X,
  pred_fun = function(m, X, ...) stats::predict(m, X, ...)$predictions,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

# S3 method for Learner
total_interaction(
  object,
  v,
  X,
  pred_fun = function(m, X) m$predict_newdata(X)$response,
  n_max = 300L,
  w = NULL,
  verbose = TRUE,
  normalize = TRUE,
  squared = TRUE,
  eps = 1e-08,
  ...
)

# S3 method for interact
total_interaction(object, normalize = TRUE, squared = TRUE, eps = 1e-08, ...)

Arguments

object

Fitted model object.

...

Additional arguments passed to pred_fun(object, X, ...), for instance type = "response" in a glm() model.

v

Vector of feature names.

X

A data.frame or matrix serving as background dataset.

pred_fun

Prediction function of the form function(object, X, ...), providing K >= 1 numeric predictions per row. Its first argument represents the model object, its second argument a data structure like X. Additional arguments (such as type = "response" in a GLM) can be passed via .... The default, stats::predict(), will work in most cases. Note that column names in a resulting matrix of predictions will be used as default column names in the results.

n_max

If X has more than n_max rows, a random sample of n_max rows is selected from X. In this case, set a random seed for reproducibility.

w

Optional vector of case weights for each row of X.

verbose

Should a progress bar be shown? The default is TRUE.

normalize

Should statistic be normalized? Default is TRUE.

squared

Should squared statistics be returned? Default is TRUE.

eps

Threshold below which numerator values are set to 0.

Value

Vector of total interaction strength (one value per prediction dimension).

Methods (by class)

total_interaction(default): Default method of total interaction strength.

total_interaction(ranger): Total interaction strength from "ranger" models.

total_interaction(Learner): Total interaction strength from "mlr3" models.

total_interaction(interact): Total interaction strength from "interact" object.

References

Friedman, Jerome H., and Bogdan E. Popescu. "Predictive Learning via Rule Ensembles." The Annals of Applied Statistics 2, no. 3 (2008): 916-54.

Examples

# MODEL ONE: Linear regression
fit <- lm(Sepal.Length ~ . + Petal.Width:Species, data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
total_interaction(inter)
#>         y 
#> 0.0502364 

if (FALSE) {
total_interaction(fit, v = names(iris[-1]), X = iris, verbose = FALSE)

# MODEL TWO: Multi-response linear regression
fit <- lm(as.matrix(iris[1:2]) ~ Petal.Length + Petal.Width * Species, data = iris)
v <- c("Petal.Length", "Petal.Width", "Species")
inter <- interact(fit, v = v, X = iris, verbose = FALSE)
total_interaction(inter)
total_interaction(fit, v = v, X = iris, verbose = FALSE)

# MODEL THREE: Linear regression
fit <- lm(Sepal.Length ~ ., data = iris)
inter <- interact(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
total_interaction(inter)
total_interaction(fit, v = names(iris[-1]), X = iris, verbose = FALSE)
}