Calculate the Transient Impression after an Event

Calculate the Transient Impression after an Event

transient_impression(act, beh, obj, dictionary = "us")

Arguments

beh
lowercase string corresponding to the behavior term

dictionary
which dictionary to use, currently set to "us"

actor
lowercase string corresponding to the actor identity

object
lowercase string corresponding to the object identity

Value

dataframe in long format, with one row for each element-dimension of the event, columns for fundamental sentiment and transient impression.

Examples

transient_impression("ceo", "advise", "benefactor")

#>         term element dimension fundamental_sentiment trans_imp
#> 1        ceo       A         E                  0.71  1.512329
#> 2        ceo       A         P                  3.22  2.377649
#> 3        ceo       A         A                  1.48  1.024836
#> 4     advise       B         E                  2.57  1.600150
#> 5     advise       B         P                  2.28  2.086134
#> 6     advise       B         A                  0.28  0.612600
#> 7 benefactor       O         E                  1.97  1.499909
#> 8 benefactor       O         P                  1.98  1.076825
#> 9 benefactor       O         A                  0.10  0.097500