Identify conditions (background, time of stimulation, biological and technical replicates) from column names

Identify conditions (background, time of stimulation, biological and technical replicates) from column names

identify_conditions(df, Column_intensity_pattern = "^Intensity.",
  split = "_", bckg_pos = 1, bio_pos = 3, time_pos = 2,
  tech_pos = 4)

Arguments

df
A dataframe containing protein intensities. By default, protein intensity column names start by "Intensity." (use parameter Column_intensity_pattern to change)

Column_intensity_pattern
Pattern (regular exrpression) used to identfy df's columns containing protein intensity values

split
Character used to split column names into substrings

bckg_pos
Position of the sample background in splitted column names

bio_pos
Position of the sample biological replicate in splitted column names

time_pos
Position of the sample experimental condition in splitted column names

tech_pos
Position of the sample technical replicate in splitted column names

Value

a data frame describing experimental samples in terms of background, biological and technical replicates, and experimental conditions

Examples

# NOT RUN {
#load data :
data("proteinGroups_Cbl")
# You can identify columns and their description separately using \code{identify_conditions()}
cond <- identify_conditions(proteinGroups_Cbl)
print.data.frame(cond)
# and use it as parameters for function InteRact()
res <- InteRact(proteinGroups_Cbl, bait_gene_name = "Cbl", condition = cond)
# }