maximally_confirm_behavior

Identify the behavior that would maximally confirm the identities of actor and object pairing

maximally_confirm_behavior(actor, object, dictionary, equation, eq_df = NULL)

Arguments

actor
lowercase string corresponding to the actor identity

object
lowercase string corresponding to the object identity

dictionary
which dictionary to use, currently set to "us"

equation
which equation to use - you can either set it to "us" for the us 1978 equations, or "user supplied")

eq_df
if you select "user supplied" for equation, this parameter should be your equation dataframe, which (should have been reshaped by the reshape_new_equation function prior)

beh
lowercase string corresponding to the behavior term

which
indicate whether you want the optimal behavior for "actor", "object", or "both"

Value

3 digit EPA indicating the optimal behavior